flask-api/rag_project/requirements.txt

langchain_community==0.1.0
torch==1.10.0
sentence-transformers==2.0.0
pymongo==4.0.1
langchain==0.1.0